+++ Sources/Core/Network/Client+Error.swift
@@ -10,9 +10,9 @@ import Alamofire
 /// Potential extra errors returned by `Client.extract(from:error)`
 public enum ClientError: Error, LocalizedError {
 	/// Message from response
-	case message(String)
+	case message(String, underlyingError: Error)
 	/// In case of HTTP Status 401 or 403
-	case unauthorized
+	case unauthorized(underlyingError: Error)
 
 	public var errorDescription: String? {
 		switch self {
@@ -20,6 +20,14 @@ public enum ClientError: Error, LocalizedError {
 			return message
 		case .unauthorized:
 			return L10n.Client.Error.unauthorized
+		}
+	}
+
+	/// The underlying error as reported by serializers, Alamofire, ...
+	public var underlyingError: Error {
+		switch self {
+		case .message(_, let error), .unauthorized(let error):
+			return error
 		}
 	}
 }
@@ -41,20 +49,20 @@ public extension Client {
 	/// - returns: An error with the message from the response (see `ClientError`), or the existing error
 	static func extract<T>(from response: DataResponse<T>, error: Error) -> Error {
 		if let status = response.response?.statusCode, status == 401 || status == 403 {
-			return ClientError.unauthorized
+			return ClientError.unauthorized(underlyingError: error)
 		}
 
 		guard let data = response.data else { return error }
 		let json = try? JSONSerialization.jsonObject(with: data, options: [])
 
-		if let json = json as? [[String: String]],
-			let message = json.first?[Keys.message] {
-			return ClientError.message(message)
-		} else if let json = json as? [String: String],
-			let message = json[Keys.message] {
-			return ClientError.message(message)
-		} else if let message = String(data: data, encoding: .utf8) {
-			return ClientError.message(message)
+		if let json = json as? [[String: Any]],
+			let message = json.first?[Keys.message] as? String {
+			return ClientError.message(message, underlyingError: error)
+		} else if let json = json as? [String: Any],
+			let message = json[Keys.message] as? String {
+			return ClientError.message(message, underlyingError: error)
+		} else if let message = String(data: data, encoding: .utf8), !message.isEmpty {
+			return ClientError.message(message, underlyingError: error)
 		} else {
 			return error
 		}